[JBPM-4304] Changing package name
