+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/main/java/org/kie/kogito/serverless/workflow/parser/DefaultNodeIdGenerator.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.kogito.serverless.workflow.parser;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+public class DefaultNodeIdGenerator implements NodeIdGenerator {
+
+    private DefaultNodeIdGenerator() {
+    }
+
+    private AtomicLong idCounter = new AtomicLong(1);
+
+    public static DefaultNodeIdGenerator get() {
+        return new DefaultNodeIdGenerator();
+    }
+
+    @Override
+    public long getId() {
+        return idCounter.getAndIncrement();
+    }
+
+}
